--- src/state/publish/get-dimension-map.js
@@ -1,11 +1,0 @@
-// @flow
-import type {
-  DimensionMap,
-    Publish,
-} from '../../types';
-export default ({
-  existing: DimensionMap,
-  publish: Publish
-}): DimensionMap => {
-
-});
